Revert changes to utils (adding get_specific_listings from whatson)
